+++ Photo Club Hub/Extensions/CoreLocation+Equatable.swift
@@ -7,7 +7,8 @@
 
 import CoreLocation
 
-extension CLLocationCoordinate2D: Equatable {
+extension CLLocationCoordinate2D: @retroactive Equatable {
+    // https://github.com/swiftlang/swift-evolution/blob/main/proposals/0364-retroactive-conformance-warning.md
 
     public static func == (lhs: CLLocationCoordinate2D, rhs: CLLocationCoordinate2D) -> Bool {
         (lhs.longitude == rhs.longitude) && (lhs.latitude == rhs.latitude)